+++ frontend/src/components/Banner.tsx
