--- frontend/src/components/MainPage/TopBar/Logo.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-//import LogoImage from '../../assets/image/LogoImage.png' //이미지 불러오기 위함
-
-function Logo() {
-    return(
-    <div>
-        {/* <img src={LogoImage} className="Logo"/> */}
-        <div className="">Grapy</div>
-    </div>
-    );
-};
-export default Logo;
