--- frontend/src/components/MainPage/TopBar/MypageIcon.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import FaceIcon from '@mui/icons-material/Face';
-
-function MypageIcon() {
-    return(
-    <div>
-        <button>
-            <FaceIcon />
-        </button>
-    </div>
-    );
-};
-export default MypageIcon;
