--- frontend/src/components/MainPage/TopBar/TopBar.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-
-const TopBar = () => {
-  //return <div className="text-blue-600">메인페이지 입니다</div>;
-  return(
-    <div>
-
-    </div>);
-};
-
-export default TopBar;
