+++ frontend/src/pages/MainPage.tsx
@@ -1,35 +1,34 @@
-import React from "react";
+import React from 'react';
 
-import TopBar from "../components/MainPage/TopBar/TopBar";
-import Banner from "../components/MainPage/Banner";
-import ProjectCard from "../components/MainPage/ProjectCard";
-
+import TopBar from '../components/TopBar';
+import Banner from '../components/Banner';
+import ProjectCard from '../components/ProjectCard';
 
 const MainPage = () => {
-  //return <div className="text-blue-600">메인페이지 입니다</div>;
-  return(
+  return (
     <div>
-      <TopBar/>
-      <Banner/>
-      <div className="justify-center content-center grid grid-cols-4 gap-4 mb-8">
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
-        <ProjectCard/>
+      {/* <TopBar/> */}
+      <Banner />
+      <div className="justify-center content-center grid grid-cols-4 gap-10 mb-8">
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
+        <ProjectCard />
       </div>
-    </div>);
+    </div>
+  );
 };
 
 export default MainPage;
